Add files via upload
just some stuff i did for roblox. i was very interested in security in programming, and one of my friends told me, in order to stop attacks, first you have to learn how they operate. so i made a simple account creation bot for the popular roblox.com. i have not and try not to use it maliciously. strictly for educational purposes. i also made a small launcher for roblox so it will automatically boot the game for me, without me having to visit the website.